# api.yaml
openapi: 3.1.0
...
info:
  title: Chatbot Administrasi Mahasiswa
  version: 0.1.0
  description: |
    API chatbot RAG untuk pertanyaan administrasi akademik mahasiswa.

    ## Lima jenis balasan

    Endpoint chat selalu mengembalikan 200 dengan `kind` yang menentukan cara
    frontend menampilkannya. `answer`, `refusal`, dan `support` WAJIB terlihat
    berbeda satu sama lain:

    | `kind` | Arti | Tampilan |
    |---|---|---|
    | `answer` | Jawaban bersumber dokumen resmi | Bubble normal + kartu sitasi (FE-2) |
    | `refusal` | Retrieval terlalu lemah, LLM tidak dipanggil (FR-3) | Visual berbeda + kontak (FE-4) |
    | `support` | Pertanyaan bernuansa tekanan mental (FR-7) | Nada empatik + kontak konseling |
    | `smalltalk` | Sapaan atau basa-basi; retrieval dan LLM dilewati | Bubble biasa, tanpa sitasi dan tanpa tombol umpan balik |
    | `rejected` | Dihentikan gerbang JEV: nonsense, upaya manipulasi, atau di luar topik kampus | Bubble biasa, tanpa sitasi dan tanpa tombol umpan balik |

    Hanya `answer` yang membawa sitasi. Menampilkan kartu sitasi di layar
    penolakan memberi kesan keliru bahwa jawabannya bersumber; menampilkannya
    pada sapaan sama saja mengklaim "halo" ada di panduan akademik.

    `smalltalk` ada supaya mahasiswa yang mengetik "hai" tidak dibalas
    penolakan berisi alamat loket biro, dan supaya sapaan tidak masuk daftar
    pertanyaan tak terjawab (AD-4).

    `rejected` hanya muncul bila gerbang JEV dinyalakan (`JEV_ENABLED=true`).
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/admin/documents:
    get:
      tags: [admin-documents]
      summary: Daftar dokumen
      operationId: list_documents
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        AD-2. Setiap baris membawa `stale` untuk badge peringatan: dokumen yang
        lebih dari 6 bulan tidak diperbarui, atau sudah lewat `valid_until`.

        PRD §12 menyebut "dokumen tidak diupdate" sebagai risiko dengan dampak
        sistem menyebarkan info kedaluwarsa secara aktif. Badge ini mitigasinya,
        jadi jangan sembunyikan di balik filter.

        Staf/dosen hanya melihat dokumen unitnya; `total` dan `stale_count` pun
        dihitung atas unit itu. Membuka, mengubah, menghapus, atau mengunggah
        dokumen unit lain dijawab 403. Nama unit dibandingkan tanpa peduli huruf
        besar dan spasi berlebih.
      parameters:
        - name: include_inactive
          in: query
          schema: { type: boolean, default: false }
        - name: only_stale
          in: query
          description: Saring hanya dokumen yang perlu ditinjau.
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Daftar dokumen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [admin-documents]
      summary: Unggah dokumen baru
      operationId: upload_document
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        AD-3. Drag & drop + form metadata. Pemrosesan berjalan sinkron: unggah,
        ekstraksi, chunking, embedding, simpan -- semuanya dalam satu transaksi.
        Dokumen yang gagal di tengah jalan tidak meninggalkan chunk yatim.

        PDF tanpa lapisan teks (hasil scan, atau dicetak dengan Print to PDF yang
        mengubah huruf menjadi garis gambar) ditolak dengan 422 (FR-1). Ini bukan
        kerewelan: chunk kosong tetap masuk indeks dan tidak pernah terambil,
        sehingga dokumen tampak terpasang padahal tidak berfungsi.

        Bahasa pesan galat ke admin harus non-teknis (PRD §9).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUpload'
            encoding:
              file:
                contentType: application/pdf
      responses:
        '201':
          description: Dokumen terpasang dan terindeks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '502':
          description: |
            Layanan luar gagal: penyimpanan dokumen atau API AI. `detail` berisi
            kalimat siap tampil yang menyarankan mencoba lagi; rinciannya hanya
            ada di log server.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: Berkas melebihi `MAX_UPLOAD_MB`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: |
            Berkas ditolak. `detail` sudah berupa kalimat siap tampil untuk admin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                pdf_scan:
                  summary: PDF tanpa lapisan teks -- hasil scan atau Print to PDF (FR-1)
                  value:
                    detail: >-
                      'Panduan Akademik 2025.pdf' tidak memiliki lapisan teks (hasil
                      scan atau dicetak dengan Print to PDF); 85% halamannya kosong.
                      Simpan ulang dari aplikasi aslinya dengan Save as PDF, atau
                      jalankan OCR bila dokumennya hasil scan.
                empty:
                  summary: Tidak menghasilkan satu chunk pun
                  value:
                    detail: "'Lampiran.pdf' tidak menghasilkan satu chunk pun"
  /api/admin/documents/{document_id}:
    parameters:
      - name: document_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [admin-documents]
      summary: Detail satu dokumen
      operationId: get_document
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Detail dokumen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [admin-documents]
      summary: Perbarui metadata atau status aktif
      operationId: update_document
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        Mengubah `valid_until` atau `is_active` langsung mempengaruhi retrieval
        pada permintaan berikutnya -- filter dokumen aktif dievaluasi di dalam
        WHERE setiap query (FR-2), bukan di-cache.

        Menonaktifkan dokumen TIDAK menghapus chunk-nya, sehingga dapat
        dikembalikan bila ternyata keliru.

        `updated_at` diperbarui saat `title`, `unit`, `effective_year`, atau
        `valid_until` berubah -- itu dianggap peninjauan dan menghapus badge
        usia AD-2. Mengubah `is_active` saja tidak: menyalakan ulang dokumen
        lama tidak boleh menghapus peringatannya.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentUpdate' }
      responses:
        '200':
          description: Tersimpan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

    delete:
      tags: [admin-documents]
      summary: Hapus dokumen beserta seluruh chunk-nya
      operationId: delete_document
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        Permanen dan tidak dapat dibatalkan. Frontend WAJIB meminta konfirmasi
        (PRD §9: semua aksi destruktif butuh konfirmasi).

        Untuk sekadar mengeluarkan dokumen dari indeks, pakai
        `PATCH { "is_active": false }` -- itu dapat dikembalikan.
      responses:
        '204':
          description: Terhapus
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/admin/documents/{document_id}/chunks:
    get:
      tags: [admin-documents]
      summary: Pratinjau hasil chunking
      operationId: list_document_chunks
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        AD-3 menyebut "preview hasil chunk". Gunanya konkret: admin dapat
        melihat apakah tabel atau daftar bernomor terpotong di tengah, yang
        merupakan penyebab paling umum jawaban benar tetapi tidak lengkap.
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Daftar chunk terurut
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Chunk' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/admin/units:
    get:
      tags: [admin-units]
      summary: Daftar semua unit layanan
      operationId: list_admin_units
      x-status: implemented
      x-min-role: superadmin
      security: [{ bearerAuth: [] }]
      description: |
        Termasuk unit nonaktif, dalam urutan menu, beserta jumlah dokumen dan
        akun yang memakainya. Untuk isian unit gunakan `GET /api/units`, yang
        hanya memuat unit aktif.
      responses:
        '200':
          description: Semua unit
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminUnit' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [admin-units]
      summary: Tambah unit layanan
      operationId: create_unit
      x-status: implemented
      x-min-role: superadmin
      security: [{ bearerAuth: [] }]
      description: |
        Unit baru langsung aktif: tampil di menu chatbot dan dapat dipilih untuk
        dokumen, entri tanya jawab, dan akun staf. `sort_order` kosong = paling akhir.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUnitCreate' }
      responses:
        '201':
          description: Unit dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUnit' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Nama sama dengan unit lain (huruf besar-kecil dan spasi tidak dibedakan)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token dari `POST /api/admin/login`. Hanya untuk endpoint admin.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    ValidationError:
      description: Body permintaan tidak sah
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HTTPValidationError' }
...
    Unauthorized:
      description: |
        Token tidak ada, tidak valid, atau kedaluwarsa; akun sudah dinonaktifkan
        atau dihapus; atau kata sandi sudah diganti sejak token terbit. Frontend
        menghapus sesi dan kembali ke halaman login.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        Level akun di bawah `x-min-role` operasi ini, atau staf/dosen menyentuh
        dokumen unit lain. Sesi tetap berlaku; `detail` siap tampil.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Sumber daya tidak ditemukan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Document:
      type: object
      required: [id, title, unit, updated_at, is_active, chunk_count, stale]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        unit: { type: string }
        effective_year: { type: [integer, 'null'] }
        valid_until:
          type: [string, 'null']
          format: date
          description: |
            null berarti berlaku tanpa batas. Lewat tanggal ini, dokumen
            berhenti terambil retrieval secara otomatis (FR-2) -- tidak perlu
            ada yang ingat menonaktifkannya.
        updated_at: { type: string, format: date-time }
        is_active: { type: boolean }
        chunk_count: { type: integer, minimum: 0 }
        stale:
          type: boolean
          description: |
            Badge peringatan AD-2: lebih dari 6 bulan tidak diperbarui, atau
            sudah lewat `valid_until`.
        uploaded_by:
          type: [string, 'null']
          description: |
            Email admin pengunggah. PRD §12 memitigasi dokumen usang lewat admin
            bernama -- badge tanpa nama penanggung jawab mudah diabaikan.
    DocumentPage:
      type: object
      required: [items, total, stale_count]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Document' }
        total:
          type: integer
          minimum: 0
          description: Jumlah dokumen yang cocok dengan filter, untuk paginasi.
        stale_count:
          type: integer
          minimum: 0
          description: |
            Dokumen AKTIF yang perlu ditinjau, tidak terpengaruh filter. Untuk
            lencana angka di navigasi AD-2. Dokumen nonaktif tidak dihitung
            karena sudah tidak terambil retrieval.
    DocumentUpload:
      type: object
      required: [file, title, unit]
      properties:
        file: { type: string, format: binary, description: Berkas PDF. }
        title: { type: string, minLength: 3, maxLength: 500 }
        unit:
          type: string
          minLength: 2
          maxLength: 200
          description: |
            Salah satu `name` dari `GET /api/units`, tanpa peduli huruf besar.
            Yang tersimpan selalu ejaan resmi; nama lain dijawab 422.
        effective_year: { type: [integer, 'null'], minimum: 2000, maximum: 2100 }
        valid_until: { type: [string, 'null'], format: date }
    DocumentUpdate:
      type: object
      description: Hanya field yang dikirim yang diubah.
      properties:
        title: { type: string, minLength: 3, maxLength: 500 }
        unit:
          type: string
          minLength: 2
          maxLength: 200
          description: |
            Salah satu `name` dari `GET /api/units`, tanpa peduli huruf besar.
            Yang tersimpan selalu ejaan resmi; nama lain dijawab 422.
        effective_year: { type: [integer, 'null'], minimum: 2000, maximum: 2100 }
        valid_until:
          type: [string, 'null']
          format: date
          description: Kirim null untuk menjadikan dokumen berlaku tanpa batas.
        is_active: { type: boolean }
    IngestionResult:
      type: object
      required: [document_id, page_count, chunk_count]
      properties:
        document_id: { type: string, format: uuid }
        page_count: { type: integer, minimum: 1 }
        chunk_count: { type: integer, minimum: 1 }
        warnings:
          type: array
          items: { type: string }
          description: |
            Catatan mutu dokumen yang baru diunggah -- mis. teks yang terbaca
            sangat sedikit karena isinya didominasi tangkapan layar. Bukan
            galat: unggahan tetap berhasil. Tiap entri berupa kalimat siap
            tampil untuk admin.
    Chunk:
      type: object
      required: [id, content, page, position]
      properties:
        id: { type: string, format: uuid }
        content: { type: string }
        page: { type: integer, minimum: 1 }
        position: { type: integer, minimum: 0 }
...
    AdminUnit:
      type: object
      required: [name, sort_order, is_active, document_count, account_count]
      description: Satu unit layanan, termasuk yang nonaktif.
      properties:
        name:
          type: string
          description: Kunci utama; nilai yang tersimpan di `documents.unit` dan `admins.unit`.
        description:
          type: [string, 'null']
          description: Teks bantu di menu chatbot.
        sort_order: { type: integer, description: Urutan tampil di menu; kecil lebih dulu. }
        is_active:
          type: boolean
          description: false = tersembunyi dari menu chatbot dan pilihan isian baru.
        document_count: { type: integer }
        account_count: { type: integer }
    AdminUnitCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Spasi berlebih dirapikan; tidak boleh memuat `/`.
        description: { type: [string, 'null'], maxLength: 500 }
        sort_order:
          type: [integer, 'null']
          minimum: 0
          maximum: 10000
          description: Kosong = diletakkan paling akhir.
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: |
            Kalimat siap tampil. Untuk endpoint admin bahasanya non-teknis
            (PRD §9: "dokumen sedang diproses", bukan "embedding in progress").
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            type: [string, integer]
        msg: { type: string }
        type: { type: string }
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items: { $ref: '#/components/schemas/ValidationError' }